worker: add header-based authenticator
With the following `[authenticator]` configuration,

    [authenticator]
    type = "header"
      [authenticator.header]
      name = "X-Remote-User"

worker will _trust_ this header, and take its value as the authenticated
user's ID.

An example use case of this feature is a service setup where an upstream
service already checks authentication, and passed that along in a
request header. Of course, incoming requests (edge of the service
network) will need to have that header stripped.

Real-world examples are Kubernetes' _Authenticating Proxy_, see here:
https://kubernetes.io/docs/admin/authentication/#authenticating-proxy

Signed-off-by: Stephan Renatus <srenatus@chef.io>