[fork] add post-run hook for setting up x-automate-authn service user
This is super-basic, and errors are ignored*. But it's enough to stand up
foulkon-worker in chef/a2, and have authn-proxy work with it.

(*) hab-sup will attempt to run hooks/post-run when the service has
been started; and if it fails, doesn't bother. So,

a) we have to take care of waiting for the service to be responsive, and
b) we have to keep in mind what none of the stuff in the hook is
   guaranteed to have happened.

   Therefore, this doesn't solve https://github.com/chef/authz-service/issues/8,
   but merely makes this container a little more useful for chef/a2.

Signed-off-by: Stephan Renatus <srenatus@chef.io>