[fork] add a way to start foulkon-proxy
I'm not sure if this is the right way to do it, but it works. If the
config key 'proxy' is set to true, this will start foulkon-proxy.

The init hook is required by both (and running it twice isn't harmful,
just ugly); the post-run will only be executed on foulkon-proxy.

Signed-off-by: Stephan Renatus <srenatus@chef.io>